fix(lab7): Правки к 7 лабе
- Убрал this
- Добавил блоки к case
- Исправил магические числа
- Исправил именование констант
- Добавил комментарии к элементам Enumeration
- Обновил UML-диграмму, перешел с drawio на EA

+++ ObjectOrientedPractics/ObjectOrientedPractics/Model/IDiscount.cs
@@ -17,14 +17,14 @@ namespace ObjectOrientedPractics.Model.Discounts
         /// </summary>
         /// <param name="items">Список товаров.</param>
         /// <returns>Размер скидки.</returns>
-        double Calculate(List<Item> items);
+        decimal Calculate(List<Item> items);
 
         /// <summary>
         /// Применяет скидку, доступную для списка товаров.
         /// </summary>
         /// <param name="items">Список товаров.</param>
         /// <returns>Размер скидки.</returns>
-        double Apply(List<Item> items);
+        decimal Apply(List<Item> items);
 
         /// <summary>
         /// Обновляет скидку на основе полученного списка товаров.
